+++ src/main/java/org/example/studybot/model/Command.java
@@ -1,4 +1,4 @@
-package org.example.studybot;
+package org.example.studybot.model;
 
 @FunctionalInterface
 public interface Command {
