+++ src/main/java/org/example/studybot/service/UserService.java
@@ -1,4 +1,7 @@
 package org.example.studybot.service;
 
+import org.springframework.stereotype.Service;
+
+@Service
 public class UserService {
 }
